requirements.txt
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
matplotlib==3.8.2
